코인오류 해결
단위수 조정 필요...

+++ Assets/MainGameBackUp/Scripts/CounterText.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public static class CounterText
 {
-    public static string UpdateText(decimal value)
+    public static string UpdateText(long value)
     {
         string number = value.ToString();
         int numDigits = number.Length;
